Portkey/examples/requirements.txt
requests
python-dotenv
zscaler-sdk-python
